Adding: Admin Menu Manager - 0.0.1
I'm not sure I can even call this a menu manager but it works, it's based off the page builder.

At the moment it creates menus, those are like your Sidebar, top nav.
Supports links in those menus.

It doesn't need the lib but I've added it enyway for now