profiler_template
There's the original and the one from BF, there's a bit more information regarding the modules that'd be nice to see along with some of the default PHP class debugging functions. We also need to add the MS for a nice looking time.
Other than that, the bar is kinda cool