CI3 Updates & Fixes 3.2.1 ( Unofficial )
Part of me thought it would be fun to update this project a little more, even though I shouldn't. Goes without saying, probably shouldn't use this out in the wild.

Note: A lot of this work was pulled from the community, anything you see out of place was probably me lol.
RIP CI3